use a hook instead
